requirements-advanced.txt
streamlit==1.28.0
ollama==0.1.7
pymupdf==1.23.8
python-docx==0.8.11
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2